Remove `version` from `WasmOptSettings`
Since we're using the `wasm-opt` library from crates.io
we'll assume that every matching version of `cargo-contract` contains the
same version of `wasm-opt`.

This doesn't hold if a user installs without the `--locked` flag though,
so we may want to add this back in the future to warn users if there are
mismatching `wasm-opt` versions.