Require env_type transcoders to be Send + Sync
They are already required to be 'static, so this seems like a reasonable
requirement. In practice, it seems only `AccountId` and `Hash`
transcoders are inserted into this collection. Adding these two traits
allows a `ContractMetadataTranscoder` to be shared between threads via
an `Arc`.  This is very useful, especially given a
`ContractMetadataTranscoder` is not even `Clone`, which makes it
unwieldy to load the metadata once and use it around the program.