fix(build): Fix installation for `ink_linting` libraries
Reuse the previous approach with hacking the workspace manifest in the
contract project.

We cannot add those libraries as direct dependencies to cargo-contract
because we use a different version of the toolchain. To make it possible
for these libraries to be found at runtime, we would also have to hack
dylint.